Implement Folia-Support

+++ src/main/java/de/rayzs/pat/utils/scheduler/PATScheduler.java
@@ -5,25 +5,25 @@ import de.rayzs.pat.utils.Reflection;
 
 public class PATScheduler {
 
-    public static PATSchedulerTask createScheduler(Runnable runnable, int time, int period) {
+    public static PATSchedulerTask createScheduler(Runnable runnable, long time, long period) {
         return Reflection.isFoliaServer()
                 ? new FoliaScheduler().getInstance(false, runnable, time, period)
                 : new BukkitScheduler().getInstance(false, runnable, time, period);
     }
 
-    public static PATSchedulerTask createScheduler(Runnable runnable, int time) {
+    public static PATSchedulerTask createScheduler(Runnable runnable, long time) {
         return Reflection.isFoliaServer()
                 ? new FoliaScheduler().getInstance(false, runnable, time)
                 : new BukkitScheduler().getInstance(false, runnable, time);
     }
 
-    public static PATSchedulerTask createAsyncScheduler(Runnable runnable, int time, int period) {
+    public static PATSchedulerTask createAsyncScheduler(Runnable runnable, long time, long period) {
         return Reflection.isFoliaServer()
                 ? new FoliaScheduler().getInstance(true, runnable, time, period)
                 : new BukkitScheduler().getInstance(true, runnable, time, period);
     }
 
-    public static PATSchedulerTask createAsyncScheduler(Runnable runnable, int time) {
+    public static PATSchedulerTask createAsyncScheduler(Runnable runnable, long time) {
         return Reflection.isFoliaServer()
                 ? new FoliaScheduler().getInstance(true, runnable, time)
                 : new BukkitScheduler().getInstance(true, runnable, time);
