+++ src/main/java/de/rayzs/pat/utils/scheduler/PATSchedulerTask.java
@@ -2,8 +2,8 @@ package de.rayzs.pat.utils.scheduler;
 
 public interface PATSchedulerTask {
 
-    PATSchedulerTask getInstance(boolean async, Runnable runnable, int time, int period);
-    PATSchedulerTask getInstance(boolean async, Runnable runnable, int time);
+    PATSchedulerTask getInstance(boolean async, Runnable runnable, long time, long period);
+    PATSchedulerTask getInstance(boolean async, Runnable runnable, long time);
 
     boolean isActive();
     void setTaskId(int taskId);
